LZ_Configuración de rutas y reorganización de las carpertas del proyecto_001
