+++ src/App.css
@@ -1,6 +1,6 @@
 .App {
   text-align: center;
-   background-image: url('../src/img/linea.png'); /*Ruta a tu imagen de fondo */
+   background-image: url('../src/assets/images/linea.png'); /*Ruta a tu imagen de fondo */
   background-size: cover; /* Asegura que la imagen cubra todo el fondo */
   background-repeat: no-repeat; /* Evita que la imagen se repita */
   background-position: center; /* Centra la imagen */
