+++ src/Components/Header/HeaderSection.tsx
@@ -3,7 +3,7 @@ import Grid from '@mui/material/Grid';
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
 import Button from '@mui/material/Button';
-import img1 from '../../img/controlL.png';
+import img1 from '../../assets/images/controlL.png';
 import ContactSection from '../Contact/ContactSection';
 
 const HeaderSectionLeft = () => {
