+++ src/Components/NavBar.tsx
@@ -1,0 +1,17 @@
+// src/components/NavBar.tsx
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+const NavBar: React.FC = () => {
+  return (
+    <nav>
+      <ul>
+        <li>
+          <Link to="/">Home</Link>
+        </li>
+      </ul>
+    </nav>
+  );
+}
+
+export default NavBar;
