+++ src/assets/images/img2.png
