+++ src/layout/Navegador/Navegador.tsx
@@ -11,7 +11,7 @@ import Avatar from '@mui/material/Avatar';
 import Button from '@mui/material/Button';
 import Tooltip from '@mui/material/Tooltip';
 import MenuItem from '@mui/material/MenuItem';
-import Logo1 from '../../img/logo.png';
+import Logo1 from '../../assets/images/logo.png';
 
 const pages = ['Populares', 'Novedades', 'Ofertas'];
 const settings = ['Perfil', 'Ajustes', 'Dashboard', 'Cerrar Sesión'];
