+++ src/pages/Home/Home.tsx
@@ -1,0 +1,30 @@
+import React from 'react';
+import '../../App.css';
+import "slick-carousel/slick/slick.css";
+import "slick-carousel/slick/slick-theme.css";
+import Navegador from '../../layout/Navegador/Navegador';
+import HeaderSection from '../../Components/Header/HeaderSection';
+import SliderSection from '../../Components/SliderImg/SliderSection';
+import NewsSection from '../../Components/Body/NewsSection';
+import NewGameSection from '../../Components/Body/NewGameSection';
+// import Api from './layout/ExampleApi/tabla';
+
+// import { Box } from '@mui/material';
+
+const Home: React.FC = () => {
+  return (
+    <div className="App">
+      <div>
+        <Navegador />
+        <HeaderSection />
+        <SliderSection />
+        <NewsSection />
+        <NewGameSection />
+        {/* <Api /> */}
+      </div>
+    </div>
+  );
+}
+
+
+export default Home;
